Separation of code into clearly separable libraries
libmlist and libsample are now completely separated from the execution
directory. Submodules automatically built and installed as either .so or .a
libraries depending on environment variable.

Turned out to be quite painful to do this transfer and there are some quirks
still, but at least code builds and runs.

From a maintenance perspective, one should perhaps adopt the Android approach
of heavily importing pieces of makefiles, but I'm not happy about it as I
don't like the Android build-system one bit. Another approach to set-up a
full-blown fool-proof build-system would be to copy Tinkers, but that would
be quite a lot of work.

A combination of both - maybe. This wasn't intend to be such a big
project. Besides, it's meant to run on Android in the end. Will see (TBD)

Problems remaining:
* make uninstall - doesn't work
* More than one header-file as public in each sub-modules wont install
* properly

Category: development