Abandoning the idea of list ID
Replacing it with handler_t type

Pointer is an excellent key as is and as ID only was meant to find lists
easily, we were in practice inventing a memory allocator. No need as there
exists an excellent one already (malloc). We just hide the pointer to avoid
encouraging direct access.

Thinking of scrapping most of the mlib API. It can't work as to be
transparent, data has to be unknown in C (void*). Hence all the "new"
operators can be scrapped as they can't be sorted and it turns out using
them is unnatural.

add_last and add_first also feels strange if add is supposed to be sorted.
It could make sense if compare function is missing I suppose, but should
lead to deliberate error.

Instead one would like to add new functions:
lseek(offset, from-position)
search

Category: development.