Replacing all assert with assert_ext
Bad habit of assuming NDEBUG will never be set is corrected. As Android
build-system setd NDEBUG default it's better to always use assert_ext when
in need of writing a speedy error-handler and to replace assert_ext with
either assert or other types as we go instead of the other way around.

assert_ext will behave like assert when NDEBUG is unset, but will exit with
error when set. Main thing is that whats inside clause will *always* be
executed (that applies to all assert_np.h versions of the macro, i.e. you
can-not miss executing code by misstake).