RT abilities
RT abilities implemented in all threads that can affect accuracy and
distribution. Both accuracy and sigma improved by a factor x2. I.e.
mean-value period time is 0.5% off at 1KHz and 95% of all are withing 10
off. This I believe is as good as it gets outside kernel and/ore bare metal.

I.e. not very impressive, unless considered that Linux is not a RTOS.