Added pre-recored data-files and harvested data-files
Restructure examples

Preparing for real development for inotify using HEMUL

Be more permissive in filenames

* Allow more characters in filenames.
* More leads to help track down syntax errors in sig-def files

Battery events signal-definition file added

File monitoring event mechanism deployed (iNotify)

Corner cases not verified, i.e. reading buffered output containing multiple
or fragmented lines of valid data.

Considering the above it might be a good idea to parse time-stamps from
logs. It's not a good idea letting them represent the sample however,
because that time-stamp is when data is read and if combined with several
sources (especially high-res timer events) we'll definitely be on thin ice.
But having them parsed might be a compromise as when the data is created is
from an event perspective more valid (for that particular data). This can
equally well be done host-side post-process however where additional logic
can also more easy and flexibly be deployed.

Logic added to keep-up reading of file

Various scenarios exist and plenty of race opportunities. This is an attempt
to keep up when several valid data-lines arrive (or has arrived already) for
the same iNotify event.

This is a common especially when sampler starts and reads from a non-empty
buffer. Things to be aware of/challenges:

* A new iNotify event occurs while still consuming dept since the last one.

* In (especially) cases when either lots of valid lines occur during the
  same notify, or when a output line is stuck in buffer, sampler time can
  diverge from the time-stamp commonly stored in the log. There's not much
  that can be done about this as this is an effect/feature from the content
  provider. But at least an optional field should be added to output a
  time stamp-files from log in parallel in a separate column.

Makefiles handle dep-files

Make-dependencies for header-files

Cherry-picked from HEMUL

Deploying new macro DBG_INF

This replaces INFO that is used only for verbose output from now on

Sub make files build ctags

Bug-fix

Detecting monitored files caught up changed

Central point is how fp_last is used and who owns it. Unfortunately there
are administrative changes here too (functions moved) which makes change
harder to review but which are of no logic interest at all.

Known issue: Every second point being caught up is missed. fp_last moved too
far somewhere?

Works well with new time-event thread (metronome)

Catches up perfectly. With a background metronome no samples are missed.

Known issues:
With bursts or updates, samples are missed. Two possibilities exist
whereas one is proven:

* Lines not ending with LF, i.e. flushed small buffer case (proven).
* Big buffer with many valid lines.

Don't miss samples on fractioned buffers

Costly when backlogged. Case optimization not to run m*n! times

Support added for tcp_tap's new abilities

Let others join a ongoing session by specifying TCP_TAP_NICNAME='@ANY@'
in a local .gdb.tap.sh_profile.rc. This should also make it easier to change
from native gdb to x-gdb only by changing session type.

From a code perspective however, nothing was changed.

Monitor generates too many self-events (bug-fixed)

Happened when catching up. Solution gives that one serious drawback is also
solved: Threads don't seek for additional hits while catch-up is pending
(serious because it involves a massive amount of potentially quite costly
fseeks). This solution works even in the case where file grows while
catching-up because iNotify will create an un unbalance in the number of
pending events and sig_data->backlog_events and make the threads run again
as soon as they are finished catching up.

Known issue: When new bursts arrive, sampler misses most. Catches either the
last or the first it seems, but the others are lost. Need to improve HEMUL
to be able to catch this in the debugger.

Other: Noticed that while debugging with GDB, iNotify is constantly firing.
Need either to filter this or debugging of the known issue will be hard
impossible to reproduce.

Wait with continuous read until device is ready.

There is a race between iNotify and readiness of data. This is not noticeable
if data arrives nicely one line at a time, but when getting data in chunks,
there will be only one iNotify for the complete chunk without data
guaranteed to be read to be read (at fread level i.e.). Waiting for data
required quite a long timeout, but by using select data is forced from
kernel to libc-level much faster (select delay reports 2uS) which is quite
acceptable.

Known issues:
Lines can still be lost if first line is not a valid hit (i.e.
pattern that we expect).

Read-buffer staring with no-hit line fail (fixed).

Strange logic, but due to that we don't want to sleep if only one line and
that line is a no-hit.

Improved log-ability

Time-stamp and PID added to error output. Feature conditioned only for
native builds. As this is already part of logcat it's defaulting not to
for Android builds.

CFLAGS alteration from command-line

For convenience and not to interfere with CLAGS or LOCAL_CFLAGS as
presidency is an issue from environment or command line, a third
variable is invented: BUILD_CFLAGS

This variable will get added to CFLAGS in the same way as LOCAL_CFLAGS is.

For example to inhibit time-stamp and PID generation for local builds:

make BUILD_CFLAGS=-DNO_TIME_PID_LOG

The mechanism is generic and anything can be passed to the build. If needed,
it can also be placed in the optional .env files (which you source) before
build.

Partial read not rewinding FP far back enough

More match discovery loop executes one time extra to detect EOF, I.e. FP
needs to rewind to next-but-last ftell and not the last. Case was easy to
reproduce with HEMUL set to run with a small buffer-size (say 256 bytes)
which increased likely-hood for case to occur.

Additionally removed obsolete code.

Improved error logging

* Returns can be logged (high debug-level as they happen often and not
  always that interesting).

* Logging with time-stamp and *thread-id* and not pid.

Known issues regarding logging:

Outputs race and compete about the console. This is relevant for host-builds
only as Android logging works differently. Solution: use a semaphore to
group the two print-lines (time-stamp/tid vs. the rest).

Build option should be provided to invalidate macros RETURN and logs macro DEG_INF
when using level above 3 as they steal execution time. (A feature of the
DBG_INF is that this can be done because the second argument is expanded in
two steps). Logging should work only for level CRIT, ERR, WARN & INFO (i.e.
0-3). Levels above 3 are debug logs only.

Duplicate outputs of data for iNotify chunks

Solves the case where output can be duplicated. This was due to using the
wrong variable to reposition FP in preparation for the next read. This case
could only happen when a chunk contains one valid line only, it's last
in buffer and is incomplete. This led to using a variable which contained
data too far back. Hence when next iNotify event occurred, it would start
reading anew too far back, hence duplicate output of something already
output.

Also fixed annoying self-event counter debug print-out.

Removed obsolete lines

Removed a case that can not occur

Trimmed timeout constant and elaborated comments

Need to keep constant as low as possible as time-out currently is the
primary mechanism for detecting if single-lined iNotify event belongs to
sig_sub or not.